# feat(benchmark): Complete large data test enablement with comprehensive performance improvements